Starting refactory
In current script the ip is the uid of the element, but if the element has more than
one network interface it appears in the topology as a different device.
Refactoring consists changing the uid in the device's hostname and for this i need to change
many parts of the script and logic.