Refactory and bugfixing
Deleted reverse_module. Renamed temp folder into naspy.
Now the UID of all elements is the hostname and not the ip anymore.
This prevents to display two times the same device with more than one ip address.
ElementManager in utilities.py is thread safe but the program isn't multi-threaded yet, this
implementation will come in the future.

BUG: if you show the web interface the elements will displayed without ip address or mac and
other devices like end devices (PCs, phones, and similar) are not shown in the final topology